% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all_ggplot_to_docx.R
\name{all_ggplot_to_pptx}
\alias{all_ggplot_to_pptx}
\title{all_ggplot_to_pptx}
\usage{
all_ggplot_to_pptx(out = "tous_les_graphs.pptx", open = TRUE, png = TRUE,
  dossier = "dessin")
}
\arguments{
\item{out}{nom du fichier pptx de sortie}

\item{open}{booleen doit on ouvrir le ficher en fin de procédure}

\item{png}{doit on enregistrer des png}

\item{dossier}{chemin du dossier de creation}
}
\description{
all_ggplot_to_pptx
}
\examples{
\dontrun{
all_ggplot_to_pptx()}
}

