% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_to_disj.R
\name{answer_to_truefalse}
\alias{answer_to_truefalse}
\title{retourne un vecteur de caractere sous forme de TRUE/FALSE/FALSE en fonction de la présen de modalite dnas la reposne. en respectant l'ordre des modalité}
\usage{
answer_to_truefalse(vec, les_modalites, sep = ",", keepNA = FALSE)
}
\arguments{
\item{vec}{a vector}

\item{les_modalites}{the modality to find}

\item{sep}{the searator to use}
}
\value{
a vector
}
\description{
retourne un vecteur de caractere sous forme de TRUE/FALSE/FALSE en fonction de la présen de modalite dnas la reposne. en respectant l'ordre des modalité
}
\examples{
exemple <- data.frame(X=6:10,A=c("a,c,b",'c,b',NA,'a','b'),B=1:5)
answer_to_truefalse(c("a","c"),find_all_modality(exemple$A))


}
