% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_numeric.R
\name{as.mon.numeric}
\alias{as.mon.numeric}
\title{transform a vector into numeric}
\usage{
as.mon.numeric(vec)
}
\arguments{
\item{vec}{}
}
\value{
a vector
}
\description{
transform a vector into numeric
}
