% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_numeric.R
\name{as_mon_numeric}
\alias{as_mon_numeric}
\title{Transform a vector into numeric if meaningful, even with bad decimal, space or \%}
\usage{
as_mon_numeric(vec)
}
\arguments{
\item{vec}{a vector}
}
\value{
a numeric vector
}
\description{
Transform a vector into numeric if meaningful, even with bad decimal, space or \%
}
\details{
Note that text and factors are not transformed as numeric (except FALSE, TRUE, F, T),
contrary to R default behavior with `as.numeric(factor())`
}
\examples{
as_mon_numeric(c("1", "0", "1"))
as_mon_numeric(c("1.3", "1,5", "1;6", "16\%", "17 87 "))
as_mon_numeric(c(TRUE, "A", "F"))
as_mon_numeric(c(TRUE, TRUE, FALSE))
as_mon_numeric(factor(c("toto", "tata", "toto")))
}
