% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_ascii.R
\docType{data}
\name{chars}
\alias{chars}
\title{Clean non-ASCII character
TODO Add to {thinkr}}
\format{An object of class \code{character} of length 16.}
\usage{
chars
}
\description{
Add an option to transform as ? if they want
}
\keyword{datasets}
