% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_names.R
\encoding{UTF-8}
\name{clean_levels}
\alias{clean_levels}
\title{Clean levels label}
\usage{
clean_levels(vec, verbose = TRUE)
}
\arguments{
\item{vec}{a factor}

\item{verbose}{}
}
\description{
Clean levels label
}
