% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_names.R
\encoding{UTF-8}
\name{clean_vec}
\alias{clean_vec}
\title{Clean character vector}
\usage{
clean_vec(vec, verbose = TRUE, unique = TRUE, keep_number = FALSE,
  translit = TRUE)
}
\arguments{
\item{vec}{character vector to clean}

\item{verbose}{booleen is the function verbose}

\item{unique}{booleen do we have to apply make_unique}

\item{keep_number}{booleen keep number at begining}

\item{translit}{booleen remove non ascii character}
}
\description{
Clean character vector
}
