% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TODO.R
\encoding{UTF-8}
\name{col_name}
\alias{col_name}
\title{from tidyr}
\usage{
col_name(x, default = stop("Please supply column name", call. = FALSE))
}
\arguments{
\item{default}{}
}
\description{
from tidyr
}
