% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levels_to_vec.R
\name{dput_levels}
\alias{dput_levels}
\title{return R instruction to create levels}
\usage{
dput_levels(vec)
}
\arguments{
\item{vec}{a factor or character vector}
}
\value{
a R instruction
}
\description{
return R instruction to create levels
}
\examples{
dput_levels(iris$Species)

}
