% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncol_to_excel.R
\name{excel_names}
\alias{excel_names}
\alias{ncol_to_excel}
\alias{excel_to_ncol}
\alias{excel_col}
\title{Get position or excel name of column}
\usage{
ncol_to_excel(n)

excel_to_ncol(col_name)

excel_col()
}
\arguments{
\item{n}{the column position}

\item{col_name}{the culumn name}
}
\description{
\code{ncol_to_excel} returns excel column name from a position number. \code{excel_to_ncol} returns excel column position number from a column name. \code{excel_col} returns all excel column name.
}
\examples{
ncol_to_excel(35)
excel_to_ncol("BF")
excel_col()
ncol_to_excel(1:6)
excel_to_ncol(c('AF', 'AG', 'AH'))
}
