% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_names.R
\name{excel_to_ncol}
\alias{excel_to_ncol}
\title{return excel column position number from a column name}
\usage{
excel_to_ncol(col_name)
}
\arguments{
\item{col_name}{the culum name}
}
\description{
return excel column position number from a column name
}
\examples{
excel_to_ncol("BF")
}
