% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx.plusieurs.R
\name{export_as_xlsx}
\alias{export_as_xlsx}
\title{exporter plusieur data.frame en xlsx}
\usage{
export_as_xlsx(file, ...)
}
\arguments{
\item{file}{chemin du fichier a creer}

\item{...}{les differents data.frame a sauvegarder}
}
\description{
exporter plusieur data.frame en xlsx
}
\examples{
data(iris)
iris2 <- iris3 <- iris
export_as_xlsx(file="out.xlsx",iris,iris2,iris3)
}
