% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx.plusieurs.R
\name{export_xl}
\alias{export_xl}
\title{export many data.frame into one xlsx}
\usage{
export_xl(path, ...)
}
\arguments{
\item{path}{chemin du fichier a creer}

\item{...}{les differents data.frame a sauvegarder}
}
\description{
export many data.frame into one xlsx
}
\examples{
data(iris)
iris2 <-  iris
export_xl(path=file.path(tempdir(),"out.xlsx"),iris,iris2)
}
