% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_to_disj.R
\name{find_all_modality}
\alias{find_all_modality}
\title{renvoie les modalites de réponses possibles d'un QCM contenu dans une seule case}
\usage{
find_all_modality(vec, sep = ",")
}
\arguments{
\item{vec}{a factor or character vector}
}
\description{
renvoie les modalites de réponses possibles d'un QCM contenu dans une seule case
}
\examples{

exemple <- data.frame(X=6:10,A=c("a,c,b",'c,b',NA,'a','b'),B=1:5)
find_all_modality(exemple$A)


}
