% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_name.R
\name{find_name}
\alias{find_name}
\title{find pattern in name's dataset}
\usage{
find_name(dataset, pattern)
}
\arguments{
\item{dataset}{a data.frame (or list or anything with names parameter)}

\item{pattern}{pattern we are looking for}
}
\value{
a list with position and value
}
\description{
find pattern in name's dataset
}
\examples{

find_name(iris,"Sepal")
}
