% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_excel_to_posixt.R
\name{from_excel_to_posixt}
\alias{from_excel_to_posixt}
\title{transform the excel numeric date format into POSIXct}
\usage{
from_excel_to_posixt(vec)
}
\arguments{
\item{vec}{a vector}
}
\description{
transform the excel numeric date format into POSIXct
}
