% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notin.R
\name{\%ni\%}
\alias{\%ni\%}
\title{not in}
\usage{
x \%ni\% table
}
\arguments{
\item{x}{vector or NULL: the values to be matched}

\item{table}{the values to be matched against}
}
\description{
not in
}
\examples{
"a" \%ni\% letters
"coucou" \%ni\% letters

}
