% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TODO.R
\encoding{UTF-8}
\name{gsub2}
\alias{gsub2}
\title{gsub qui conserve le type factor si présent, et qui ne touche pas au numerique}
\usage{
gsub2(x, ...)
}
\arguments{
\item{x}{un vecteur}

\item{...}{les parametres de la fonction gsub}
}
\description{
gsub qui conserve le type factor si présent, et qui ne touche pas au numerique
}
