% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TODO.R
\encoding{UTF-8}
\name{init_summary_each}
\alias{init_summary_each}
\title{met a 1 ou n le compteur de summary_each}
\usage{
init_summary_each(n = 1)
}
\description{
met a 1 ou n le compteur de summary_each
}
