% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TODO.R
\encoding{UTF-8}
\name{is.01}
\alias{is.01}
\title{renvoie vrai si le vecteur ne contient que des 1 et des 0}
\usage{
is.01(x)
}
\arguments{
\item{x}{un vecteur}
}
\description{
renvoie vrai si le vecteur ne contient que des 1 et des 0
}
