% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TODO.R
\encoding{UTF-8}
\name{is.12}
\alias{is.12}
\title{renvoie vrai si le vecteur ne contient que des 1 et des 2}
\usage{
is.12(x)
}
\arguments{
\item{x}{}
}
\description{
renvoie vrai si le vecteur ne contient que des 1 et des 2
}
