% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TODO.R
\encoding{UTF-8}
\name{is.pos}
\alias{is.pos}
\title{indique si la class de l'objet est posixct}
\usage{
is.pos(x)
}
\arguments{
\item{x}{}
}
\description{
indique si la class de l'objet est posixct
}
