% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_full_figures.R
\encoding{UTF-8}
\name{is_full_figures}
\alias{is_full_figures}
\title{Predicate for charater vector full of figures}
\usage{
is_full_figures(.)
}
\arguments{
\item{.}{a vector of character (and eventually NA's)}
}
\value{
a boolean
}
\description{
detects if a character vector is only made with figures. Useful when you
}
\examples{
is_full_figures(c(NA,"0","25.3"))
is_full_figures((c(NA,"0","25_3")))
}
