% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/is_full_na.R
\name{is_full_na}
\alias{is_full_na}
\title{Predicate for full NA vector}
\usage{
is_full_na(.)
}
\arguments{
\item{.}{a vector}
}
\value{
a vector of boolean
}
\description{
is_full_na test if the vector is full of NA's
}
\examples{
is_full_na(c(NA, NA, NA))
}

