% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_likert.R
\name{is_likert}
\alias{is_likert}
\title{is a factor a likert scale}
\usage{
is_likert(vec, lev)
}
\arguments{
\item{vec}{a factor}

\item{lev}{le scale}
}
\value{
boolean
}
\description{
is a factor a likert scale
}
\examples{
is_likert(iris$Species,c("setosa","versicolor","virginica"))
is_likert(iris$Species,c("setosa","versicolor","virginica","banana"))
is_likert(iris$Species,c("setosa","versicolor"))

}
