% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_ressemble_a_un_nombre.R
\name{look_like_a_number}
\alias{look_like_a_number}
\title{return TRUE if this look like a number}
\usage{
look_like_a_number(vec)
}
\arguments{
\item{vec}{a vector}
}
\value{
un booleen
}
\description{
return TRUE if this look like a number
}
