% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_unique.R
\encoding{UTF-8}
\name{make_unique}
\alias{make_unique}
\title{make.unique improvement}
\usage{
make_unique(vec, sep = "_")
}
\arguments{
\item{vec}{a vector}

\item{sep}{char  separator to use}
}
\value{
a vector
}
\description{
make.unique improvement
}
\examples{

make_unique(c("a","a","a","b","a","b","c"))

}
