% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_names.R
\name{ncol_to_excel}
\alias{ncol_to_excel}
\title{return excel column name from a position number}
\usage{
ncol_to_excel(n)
}
\arguments{
\item{n}{the colum position}
}
\description{
return excel column name from a position number
}
\examples{
ncol_to_excel(35)
}
