% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ublo.R
\name{peep}
\alias{peep}
\title{peep the pipeline}
\usage{
peep(data, ..., printer = print, verbose = FALSE)
}
\arguments{
\item{data}{some data}

\item{\dots}{function names or expressions that use \code{.} as a
placeholder for the data}

\item{printer}{which function use to print}

\item{verbose}{TRUE to include what is printed}
}
\value{
the input data
}
\description{
peep some data at one step of a pipeline.
}
\examples{
if( require(magrittr) ){
  # just symbols
  iris \%>\% peep(head,tail) \%>\% summary
  # expressions with .
  iris \%>\% peep(head(., n=2),tail(., n=3) ) \%>\% summary
  # or both
  iris \%>\% peep(head,tail(., n=3) ) \%>\% summary
  # use verbose to see what happens
  iris \%>\% peep(head,tail(., n=3), verbose = TRUE) \%>\% summary
}
}
