% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TODO.R
\encoding{UTF-8}
\name{replace_pattern}
\alias{replace_pattern}
\title{Replace pattern everywhere in a data.frame}
\usage{
replace_pattern(dataset, pattern, replacement, exact = FALSE)
}
\description{
Replace pattern everywhere in a data.frame
}
\examples{
library(dplyr)
library(tidyr)
dataset <-
data.frame(a=as.factor(letters)[1:7],b=letters[1:7],c=1:7,stringsAsFactors = FALSE) \%>\%
unite("fus",a,b,remove=FALSE,sep="")
dataset \%>\% replace_pattern("a",'"XXX') \%>\% summary()




}
