% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_csv.R
\encoding{UTF-8}
\name{save_as_csv}
\alias{save_as_csv}
\title{export a data.frame to csv}
\usage{
save_as_csv(dataset, path, row.names = FALSE, ...)
}
\arguments{
\item{dataset}{a data.frame}

\item{path}{the path}

\item{row.names}{booleen do we have to save the row names}

\item{...}{other write.csv parameters}
}
\value{
file name as character
}
\description{
export a data.frame to csv
}
\examples{

\dontrun{
iris \%>\% save_as_csv(file.path(tempdir(),'coucou.csv')) \%>\% browseURL()


}

}
