% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_to_disj.R
\encoding{UTF-8}
\name{separate_to_disj}
\alias{separate_to_disj}
\title{transform a variable to disjonctif table}
\usage{
separate_to_disj(dataset, col, sep = ",", keepNA = FALSE)
}
\arguments{
\item{dataset}{the data.frame}

\item{col}{the colonne to transform}

\item{sep}{the separator to split col}

\item{keepNA}{keep NA instead of FALSE}
}
\value{
a data.frame
}
\description{
transform a variable to disjonctif table
}
\examples{
exemple <- data.frame(X=6:10,A=c("a, c, b",'c,b',NA,'a','b'),B=1:5)
separate_to_disj(exemple,A,sep=",")
separate_to_disj_(exemple,"A",sep=",")


}
