% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_to_disj.R
\name{separate_to_disj_}
\alias{separate_to_disj_}
\title{transform a variable to disjonctif table}
\usage{
separate_to_disj_(dataset, col, sep = ",", keepNA = FALSE)
}
\arguments{
\item{dataset}{the data.frame}

\item{col}{the colonne to transform}

\item{sep}{the separator to split col}

\item{keepNA}{keep NA instead of FALSE}
}
\value{
a data.frame
}
\description{
transform a variable to disjonctif table
}
