% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TODO.R
\encoding{UTF-8}
\name{set_col_type}
\alias{set_col_type}
\title{applique le type précisé a chaque colonne d 'un jeu de données}
\usage{
set_col_type(dataset, col_type)
}
\arguments{
\item{col_type}{unv ecteur de chaine de caractere qui contient les différentes class}
}
\description{
applique le type précisé a chaque colonne d 'un jeu de données
}
