% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TODO.R
\encoding{UTF-8}
\name{summary_each}
\alias{summary_each}
\title{affiche les info de chaque colonne l'une apres l'autre}
\usage{
summary_each(dataset)
}
\arguments{
\item{dataset}{}
}
\description{
affiche les info de chaque colonne l'une apres l'autre
}
