% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thinkr.r
\docType{package}
\name{thinkr-package}
\alias{thinkr}
\alias{thinkr-package}
\title{thinkr: Tools for Cleaning Up Messy Files}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Some tools for cleaning up messy 'Excel' files to be suitable for R. People who have been working with 'Excel' for years built more or less complicated sheets with names, characters, formats that are not homogeneous. To be able to use them in R nowadays, we built a set of functions that will avoid the majority of importation problems and keep all the data at best.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Thinkr-open/thinkr}
  \item Report bugs at \url{https://github.com/Thinkr-open/thinkr/issues}
}

}
\author{
\strong{Maintainer}: Vincent Guyader \email{vincent@thinkr.fr} (\href{https://orcid.org/0000-0003-0671-9270}{ORCID})

Authors:
\itemize{
  \item Sébastien Rochette \email{sebastien@thinkr.fr} (\href{https://orcid.org/0000-0002-1565-9313}{ORCID})
}

Other contributors:
\itemize{
  \item ThinkR [copyright holder]
}

}
\keyword{internal}
