% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ublo.R
\name{ublo}
\alias{ublo}
\alias{ublo_}
\title{Execute some functions inside the dplyr flow}
\usage{
ublo(dataset, ...)
}
\arguments{
\item{dataset}{A tbl, a data.frame.}

\item{...}{a Comma separated list of unquoted function. use ublo_ to do standard evaluation}
}
\value{
the same dataset as input
}
\description{
allow to have a look inside the dplyr flow. see Example.
}
\examples{
data(iris)
if( require(magrittr)){
  iris \%>\% ublo(head,summary) \%>\% plot
  iris \%>\% ublo_("head","summary") \%>\% plot
}
}
