% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_ggplot_to_docx.R
\encoding{UTF-8}
\name{all_ggplot_to_pptx}
\alias{all_ggplot_to_pptx}
\title{Save all ggplot in a pptx}
\usage{
all_ggplot_to_pptx(out = "tous_les_graphs.pptx", open = TRUE, png = TRUE,
  folder = "dessin", global = TRUE)
}
\arguments{
\item{out}{output file name}

\item{open}{booleen open file after creation}

\item{png}{booleen also save as png}

\item{folder}{png's folder}

\item{global}{booleen use .GlobalEnv}
}
\description{
Save all ggplot in a pptx
}
\examples{
\dontrun{
all_ggplot_to_pptx()}
}
