% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_numeric.R
\name{as_mon_numeric}
\alias{as_mon_numeric}
\title{transform a vector into numeric}
\usage{
as_mon_numeric(vec)
}
\arguments{
\item{vec}{a vector}
}
\value{
a numeric vector
}
\description{
transform a vector into numeric
}
