% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_names.R
\encoding{UTF-8}
\name{clean_levels}
\alias{clean_levels}
\title{Clean levels label}
\usage{
clean_levels(vec, verbose = FALSE, translit = FALSE, punct = FALSE)
}
\arguments{
\item{vec}{a factor}

\item{verbose}{booleen is the function verbose}

\item{translit}{booleen remove non ascii character}

\item{punct}{booleen do you remove punctuation}
}
\description{
Clean levels label
}
