% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_names.R
\encoding{UTF-8}
\name{clean_names}
\alias{clean_names}
\title{clean_names}
\usage{
clean_names(dataset, verbose = FALSE, translit = TRUE)
}
\arguments{
\item{dataset}{a dataframe}

\item{verbose}{logical}

\item{translit}{logical remove non ascii character}
}
\value{
a dataframe
}
\description{
clean_names
}
\examples{
data(iris)
clean_names(iris)

}
