% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_names.R
\encoding{UTF-8}
\name{clean_vec}
\alias{clean_vec}
\title{Clean character vector}
\usage{
clean_vec(vec, verbose = FALSE, unique = TRUE, keep_number = FALSE,
  translit = TRUE, punct = TRUE)
}
\arguments{
\item{vec}{character vector to clean}

\item{verbose}{logical is the function verbose}

\item{unique}{logical do we have to apply make_unique}

\item{keep_number}{logical keep number at begining}

\item{translit}{logical remove non ascii character}

\item{punct}{logical do you remove punctuation}
}
\description{
Clean character vector
}
