% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_excel_to_posixt.R
\name{from_excel_to_posixt}
\alias{from_excel_to_posixt}
\title{transform the excel numeric date format into POSIXct}
\usage{
from_excel_to_posixt(vec, origin = "1904-01-01")
}
\arguments{
\item{vec}{a vector}

\item{origin}{a date-time object, or something which can be coerced by as.POSIXct(tz = "GMT") to such an object.}
}
\description{
transform the excel numeric date format into POSIXct
}
