% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_pattern.R
\encoding{UTF-8}
\name{gsub2}
\alias{gsub2}
\title{like gsub but keep a factor as factor}
\usage{
gsub2(x, ...)
}
\arguments{
\item{x}{a vector}

\item{...}{les parametres de la fonction gsub}
}
\value{
a vector
}
\description{
like gsub but keep a factor as factor
}
