% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_misc.R
\encoding{UTF-8}
\name{is.01}
\alias{is.01}
\title{does this vector only contains 0 and 1}
\usage{
is.01(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a boolean
}
\description{
does this vector only contains 0 and 1
}
\examples{

is.01(c(0,1,0,0,1))
is.01(c(0,1,0,0,5))

}
