% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_misc.R
\encoding{UTF-8}
\name{is.12}
\alias{is.12}
\title{does this vector only contains 1 and 2}
\usage{
is.12(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a boolean
}
\description{
does this vector only contains 1 and 2
}
\examples{

is.12(c(1,1,2,1,2))
is.12(c(1,1,2,1,5))

}
