% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_pattern.R
\encoding{UTF-8}
\name{replace_pattern}
\alias{replace_pattern}
\title{Replace pattern everywhere in a data.frame}
\usage{
replace_pattern(dataset, pattern, replacement, exact = FALSE)
}
\arguments{
\item{dataset}{a data.frame}

\item{pattern}{Pattern to look for.}

\item{replacement}{A character of replacements.}

\item{exact}{a boolean if TRUE the whole value need ton match}
}
\value{
a data.frame
}
\description{
Replace pattern everywhere in a data.frame
}
\examples{
library(dplyr)
library(tidyr)
dataset <-
data.frame(a=as.factor(letters)[1:7],b=letters[1:7],c=1:7,stringsAsFactors = FALSE) \%>\%
unite("fus",a,b,remove=FALSE,sep="")
dataset \%>\% replace_pattern("a",'"XXX') \%>\% summary()




}
