% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_pattern.R
\encoding{UTF-8}
\name{set_col_type}
\alias{set_col_type}
\title{set a given coltype to each column in a data.frame}
\usage{
set_col_type(dataset, col_type)
}
\arguments{
\item{dataset}{a data.frame}

\item{col_type}{a character vector containing the class to apply}
}
\value{
a data.frame
}
\description{
set a given coltype to each column in a data.frame
}
